src/data_storage.py
from google.cloud import storage
import json
from datetime import datetime
import logging

logger = logging.getLogger(__name__)

class DataStorage:
    def __init__(self, bucket_name):
        # Implementation...
        pass

    def save_videos_data(self, videos_data):
        # Implementation...
        pass

    def save_comments_data(self, video_id, comments_data):
        # Implementation...
        pass

    def load_data(self, blob_name):
        # Implementation...
        pass