requirements.txt
flask
google-cloud-storage
google-cloud-language
google-api-python-client
textblob
python-dotenv